[grid] in tlau add -grid mode to lauX.sh
(August 2024: Fix conflicts in epochX/cudacpp/tlau/lauX.sh after rebasing on upstream/master)

+++ epochX/cudacpp/tlau/lauX.sh
@@ -10,17 +10,20 @@ scrdir=$(cd $(dirname $0); pwd)
 
 function usage()
 {
-  echo "Usage:   $0 -<backend> <procdir>"
+  echo "Usage:   $0 -<backend> [-grid] <procdir>"
   echo "         (supported <backend> values: fortran, cuda, hip, cppnone, cppsse4, cppavx2, cpp512y, cpp512z)"
   echo "Example: $0 -cppavx2 gg_tt.mad"
   exit 1
 }
 
 bckend=
+grid=
 proc=
 while [ "$1" != "" ]; do
   if [ "$1" == "-fortran" ] || [ "$1" == "-cuda" ] || [ "$1" == "-hip" ] || [ "$1" == "-cppnone" ] || [ "$1" == "-cppsse4" ] || [ "$1" == "-cppavx2" ] || [ "$1" == "-cpp512y" ] || [ "$1" == "-cpp512z" ]; then
     if [ "${bckend}" == "" ]; then bckend=${1/-/}; else echo "ERROR! Backend already set"; usage; fi
+  elif [ "$1" == "-grid" ]; then
+    grid=${1}
   elif [ "${proc}" == "" ]; then
     proc=$1
   else
@@ -84,6 +87,7 @@ echo "r=21" > SubProcesses/randinit # just in case a previous test was not clean
 cp SubProcesses/randinit SubProcesses/randinit.BKP # save the initial file
 sed -i "s/.* = nevents/  10000 = nevents/" Cards/run_card.dat # just in case
 sed -i "s/.* = cudacpp_backend/ cpp = cudacpp_backend/" Cards/run_card.dat # just in case
+sed -i "s/.* = gridpack/  False = gridpack/" Cards/run_card.dat # just in case
 cp Cards/run_card.dat Cards/run_card.dat.BKP # save the initial file
 sed -i "s/      NEVENTS = .*/      NEVENTS = 10000/" Source/run_card.inc # just in case
 cp Source/run_card.inc Source/run_card.inc.BKP # save the initial file
@@ -109,6 +113,9 @@ sed -i "s/ 10000 = nevents/ ${nevt} = nevents/" Cards/run_card.dat
 # Set the backend in run_card.dat
 sed -i "s/ cpp = cudacpp_backend/${bckend} = cudacpp_backend/" Cards/run_card.dat
 
+# Set gridpack mode in run_card.dat
+if [ "${grid}" != "" ]; then sed -i "s/.* = gridpack/  True = gridpack/" Cards/run_card.dat; fi
+
 # Launch (generate_events)
 # (BUG #683: generate_events does not return an error code even if it fails)
 ###set -x # verbose
@@ -119,6 +126,9 @@ echo "END: $(date)" |& tee -a ${resultsdir}/output.txt
 END=$(date +%s)
 echo "ELAPSED: $((END-START)) seconds" |& tee -a ${resultsdir}/output.txt
 ###set +x # not verbose
+
+# Skip cleanup (for the moment?) in gridpack mode
+if [ "${grid}" != "" ]; then exit 0; fi
 
 # Process and keep results
 \rm HTML/results.pkl
